+++ tests/Model/CompanyTest.php
@@ -2,6 +2,7 @@
 
 namespace CommunityDS\Deputy\Api\Tests\Model;
 
+use CommunityDS\Deputy\Api\Model\Company;
 use CommunityDS\Deputy\Api\Tests\Adapter\MockClient;
 use CommunityDS\Deputy\Api\Tests\TestCase;
 
@@ -91,5 +92,29 @@ class CompanyTest extends TestCase
             $actual[] = $unit->id;
         }
         $this->assertEquals($expected, $actual);
+    }
+    
+    public function testSetSettings()
+    {
+        $company = $this->wrapper()->getCompany(MockClient::COMPANY_FIRST);
+    
+        $resultSetSettings = $company->setSettings([
+            Company::SETTING_WEEK_START => 1,
+            Company::SETTING_CAN_BUMP_SHIFT_VIA_DESK => true,
+        ]);
+        $this->assertTrue($resultSetSettings, 'Failed to set company settings');
+    
+        $this->assertEquals(1, $company->getSetting(Company::SETTING_WEEK_START), "Error in getSetting('{Company::SETTING_WEEK_START}')");
+        $this->assertEquals(true, $company->getSetting(Company::SETTING_CAN_BUMP_SHIFT_VIA_DESK), "Error in getSetting('{Company::SETTING_CAN_BUMP_SHIFT_VIA_DESK}')");
+    }
+    
+    public function testGetSettings()
+    {
+        $company = $this->wrapper()->getCompany(MockClient::COMPANY_FIRST);
+    
+        $this->assertTrue(is_array($company->getSettings()), 'Settings are not an array');
+    
+        $this->assertEquals(1, $company->getSetting(Company::SETTING_WEEK_START), "Error in getSetting('{Company::SETTING_WEEK_START}')");
+        $this->assertEquals(true, $company->getSetting(Company::SETTING_CAN_BUMP_SHIFT_VIA_DESK), "Error in getSetting('{Company::SETTING_CAN_BUMP_SHIFT_VIA_DESK}')");
     }
 }