+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## 1.3.0
+
+* Added ability to create `Memo` to assigned recipients ([#5](https://github.com/communityds/deputy-api-wrapper/pull/5)).
+* Added `Roster` helper method `isTimesheetCreated` ([#7](https://github.com/communityds/deputy-api-wrapper/pull/7)).
+* Added ability to change `Company` settings ([#8](https://github.com/communityds/deputy-api-wrapper/pull/8)).
+* Added support for `CustomField` and `CustomFieldData` ([#9](https://github.com/communityds/deputy-api-wrapper/pull/9)).
+* Fixed issue when updating `Roster` instances caused by missing payload information ([#6](https://github.com/communityds/deputy-api-wrapper/pull/6)).
+
 ## 1.2.0
 
 * Added PHP 7 support.